requirements.txt
aiogram==3.25.0
redis==5.0.8
python-dotenv==1.0.1
pytest==8.3.3
certifi>=2024.0.0
